requirements_web.txt
Flask==2.3.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
